Move TestdoxTerminalReporter to its own module
