+++ pytest_testdox/plugin.py
@@ -1,5 +1,5 @@
 import sys
-from typing import Generator, List, Optional, TextIO, Tuple
+from typing import Generator
 
 import pytest
 
@@ -11,10 +11,9 @@ except ImportError:  # For pytest < 7.0.0
     from _pytest.reports import TestReport
     from _pytest.runner import CallInfo
 
-from _pytest.terminal import TerminalReporter
 from pytest import Item
 
-from pytest_testdox import constants, data_structures, wrappers
+from pytest_testdox import constants, terminal
 
 
 def pytest_addoption(parser: Parser):
@@ -64,7 +63,9 @@ def pytest_configure(config: Config):
     if should_enable_plugin(config):
         # Get the standard terminal reporter plugin and replace it with ours
         standard_reporter = config.pluginmanager.getplugin('terminalreporter')
-        testdox_reporter = TestdoxTerminalReporter(standard_reporter.config)
+        testdox_reporter = terminal.TestdoxTerminalReporter(
+            standard_reporter.config
+        )
         config.pluginmanager.unregister(standard_reporter)
         config.pluginmanager.register(testdox_reporter, 'terminalreporter')
 
@@ -89,68 +90,6 @@ def pytest_runtest_makereport(
 
     if testdox_class_name:
         report.testdox_class_name = testdox_class_name
-
-
-class TestdoxTerminalReporter(TerminalReporter):  # type: ignore
-    def __init__(self, config: Config, file: TextIO = None):
-        super().__init__(config, file)
-        self._last_header_id: Optional[str] = None
-        self.pattern_config = data_structures.PatternConfig(
-            files=self.config.getini('python_files'),
-            functions=self.config.getini('python_functions'),
-            classes=self.config.getini('python_classes'),
-        )
-        self.result_wrappers: List[type] = []
-
-        if config.getini('testdox_format') != 'plaintext':
-            self.result_wrappers.append(wrappers.UTF8Wrapper)
-
-        if config.option.color != 'no':
-            self.result_wrappers.append(wrappers.ColorWrapper)
-
-    def _register_stats(self, report: TestReport):
-        """
-        This method is not created for this plugin, but it is needed in order
-        to the reporter display the tests summary at the end.
-
-        Originally from:
-        https://github.com/pytest-dev/pytest/blob/47a2a77/_pytest/terminal.py#L198-L201
-        """
-        res = self.config.hook.pytest_report_teststatus(
-            report=report, config=self.config
-        )
-        category = res[0]
-        self.stats.setdefault(category, []).append(report)
-        self._tests_ran = True
-
-    def pytest_runtest_logreport(self, report: TestReport) -> None:
-        self._register_stats(report)
-
-        if report.when != 'call' and not report.skipped:
-            return
-
-        result = data_structures.Result.create(report, self.pattern_config)
-
-        for wrapper in self.result_wrappers:
-            result = wrapper(result)
-
-        if result.header_id != self._last_header_id:
-            self._last_header_id = result.header_id
-            self._tw.sep(' ')
-            self._tw.line(result.header)
-
-        self._tw.line(str(result))
-
-    def pytest_runtest_logstart(
-        self, nodeid: str, location: Tuple[str, Optional[int], str]
-    ) -> None:
-        # Ensure that the path is printed before the
-        # 1st test of a module starts running.
-        self.write_fspath_result(nodeid, '')
-
-        # To support Pytest < 6.0.0
-        if hasattr(self, 'flush'):
-            self.flush()
 
 
 def _first(iterator):
